refactor: update ErrorUIState import
