--- app/src/main/java/com/karrar/movieapp/ui/category/uiState/ErrorUIState.kt
@@ -1,6 +1,0 @@
-package com.karrar.movieapp.ui.category.uiState
-
-data class ErrorUIState(
-    val code: Int,
-    val message: String
-)
