+++ app/src/main/java/com/karrar/movieapp/ui/movieDetails/saveMovie/uiState/MyListUIState.kt
@@ -1,6 +1,6 @@
 package com.karrar.movieapp.ui.movieDetails.saveMovie.uiState
 
-import com.karrar.movieapp.ui.category.uiState.ErrorUIState
+import com.karrar.movieapp.ui.explore.exploreUIState.ErrorUIState
 
 data class MySavedListUIState(
     val myListItemUI: List<MyListItemUI> = emptyList(),
